feat: namespacing of api sections
