+++ src/douze/api.py
@@ -1,12 +1,12 @@
+from functools import cached_property
 from logging import getLogger
 from os import getenv
-from typing import Any, Callable, Iterator, List, Optional, Text, Union
+from typing import Any, Callable, Optional, Text
 
 from httpx import HTTPError, Timeout
 from typefit import api
 from typefit import httpx_models as hm
 
-from .models import *
 from .types import UndefinedSerializer
 
 logger = getLogger("douze.api")
@@ -28,22 +28,29 @@ class DoApi(api.SyncClient):
         self.api_token = api_token
         self.helper.http.timeout = Timeout(30.0)
 
-        self._apps = None
-        self._uptime = None
-
-    @property
+    @cached_property
     def apps(self):
-        from douze.apps.api import AppsApi
+        from .apps.api import AppsApi
 
-        self._apps = self._apps or AppsApi(self)
-        return self._apps
+        return AppsApi(self)
 
-    @property
+    @cached_property
+    def droplet(self):
+        from .droplet.api import DropletApi
+
+        return DropletApi(self)
+
+    @cached_property
+    def db(self):
+        from .db.api import DatabaseApi
+
+        return DatabaseApi(self)
+
+    @cached_property
     def uptime(self):
-        from douze.checks.api import ChecksApi
+        from .uptime.api import ChecksApi
 
-        self._apps = self._apps or ChecksApi(self)
-        return self._apps
+        return ChecksApi(self)
 
     def init_serialize(self) -> Callable[[Any], Any]:
         """
@@ -152,140 +159,3 @@ class DoApi(api.SyncClient):
                 page += 1
             else:
                 break
-
-    @api.get("databases?page={page}")
-    def _db_cluster_list(self, page) -> DatabaseClusterCollection:
-        """
-        Fetches one page of database clusters list
-        """
-
-    def db_cluster_list(self) -> Iterator[DatabaseCluster]:
-        """
-        Lists all database clusters present on the account
-
-        Notes
-        -----
-        Since new versions of DB engines get added and so forth, we don't want
-        to avoid a crash in case things are not fit. That's why the databases
-        attributes also allows to not fit. Here we're checking that we're
-        returning only the instances that are indeed recognized. The other ones
-        are just ignored.
-        """
-
-        for cluster in self._iterate_collection(self._db_cluster_list, "databases"):
-            if isinstance(cluster, DatabaseCluster):
-                yield cluster
-
-    @api.post("databases", json=lambda cluster: cluster, hint="database")
-    def db_cluster_create(self, cluster: DatabaseClusterCreate) -> DatabaseCluster:
-        """
-        Creates a database cluster
-        """
-
-    @api.get("databases/{cluster_id}", hint="database")
-    def db_cluster_get(self, cluster_id: Text) -> DatabaseCluster:
-        """
-        Returns a database cluster
-        """
-
-    @api.get("databases/{cluster_id}/dbs?page={page}")
-    def _db_database_list(self, page: int, cluster_id: Text) -> DatabaseCollection:
-        """
-        Retrieves a single page of the databases list for a cluster
-        """
-
-    def db_database_list(self, cluster_id: Text) -> Iterator[Database]:
-        """
-        Iterates through all the databases present in a cluster
-        """
-
-        yield from self._iterate_collection(
-            self._db_database_list, "dbs", cluster_id=cluster_id
-        )
-
-    @api.post("databases/{cluster_id}/dbs", hint="db", json=lambda database: database)
-    def db_database_create(self, cluster_id: Text, database: Database) -> Database:
-        """
-        Creates a database within a cluster
-        """
-
-    @api.delete("databases/{cluster_id}/dbs/{database_name}")
-    def db_database_delete(self, cluster_id: Text, database_name: Text) -> None:
-        """
-        Deletes the specified database within the cluster
-        """
-
-    @api.get("databases/{cluster_id}/firewall", hint="rules")
-    def db_firewall_list(self, cluster_id: Text) -> List[DatabaseFirewallRule]:
-        """
-        Lists all firewall rules for a given cluster
-        """
-
-    @api.put("databases/{cluster_id}/firewall", json=lambda rules: {"rules": rules})
-    def db_firewall_update(
-        self,
-        cluster_id: Text,
-        rules: List[Union[DatabaseFirewallRule, DatabaseFirewallRuleCreate]],
-    ) -> None:
-        """
-        Updates the firewall rules for that cluster
-        """
-
-    @api.get("databases/{cluster_id}/users", hint="users")
-    def db_user_list(self, cluster_id: Text) -> List[DatabaseUser]:
-        """
-        Retrieves the list of users found in a cluster
-        """
-
-    @api.post("databases/{cluster_id}/users", hint="user", json=lambda user: user)
-    def db_user_create(
-        self, cluster_id: Text, user: DatabaseUserCreate
-    ) -> DatabaseUser:
-        """
-        Creates a user within a cluster
-        """
-
-    @api.delete("databases/{cluster_id}/users/{user_name}")
-    def db_user_delete(self, cluster_id: Text, user_name: Text) -> None:
-        """
-        Deletes a user within a cluster
-        """
-
-    @api.get("databases/{cluster_id}/pools", hint="pools")
-    def db_pool_list(self, cluster_id: Text) -> List[DatabaseConnectionPool]:
-        """
-        Lists all connection pools present in the cluster
-        """
-
-    @api.post("databases/{cluster_id}/pools", hint="pool", json=lambda pool: pool)
-    def db_pool_create(
-        self, cluster_id: Text, pool: DatabaseConnectionPoolCreate
-    ) -> DatabaseConnectionPool:
-        """
-        Creates a connection pool for that cluster
-        """
-
-    @api.delete("databases/{cluster_id}/pools/{pool_name}")
-    def db_pool_delete(self, cluster_id: Text, pool_name: Text) -> None:
-        """
-        Deletes a connection pool for that cluster
-        """
-
-    @api.get("database/{cluster_id}/pools/{pool_name}", hint="pool")
-    def db_pool_get(self, cluster_id: Text, pool_name: Text) -> DatabaseConnectionPool:
-        """
-        Returns a connection pool from that cluster
-        """
-
-    @api.get("droplets?page={page}")
-    def _droplet_list(self, page: int) -> DropletCollection:
-        """
-        Lists droplets on a specific page of the list
-        """
-
-    def droplet_list(self) -> Iterator[Droplet]:
-        """
-        Iterates through all the droplets in the account
-        """
-
-        yield from self._iterate_collection(self._droplet_list, "droplets")
