+++ src/douze/uptime/models.py
